Split cleanup marker atomic writer
